+++ embeddings/vertexai/vertexai_palm_test.go
@@ -1,4 +1,4 @@
-package vertexai_palm
+package vertexai
 
 import (
 	"context"
@@ -7,7 +7,6 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
 	"github.com/tmc/langchaingo/llms/vertexai"
 )
